requirements.txt
paddlepaddle
paddleocr
matplotlib
Pillow
numpy
easyocr
opencv-python
